A simple map in form of a Binary Search Tree
Task:
Implement a simple map, in the form of a binary tree, that holds an address book; the key for the map
should be a person's name and the value should be the person's email address. You should provide the
ability to add email addresses to the map, remove email addresses, update email addresses, and of
course find email addresses. You'll also want to clean up your address book when your program shuts
down. As a reminder, you can use any of the standard C++ comparison operators (such as ==, < or >) to
compare two strings.
P.S. I went a little bit further and add my own functions that do: 
- displaying the number of users in the map;
- displaying the number of users with equal names in the map;
- searching and preventing input of equal e-mails;